Move `:mode_override` to Application config
At the moment we are using `Mix.env/0` to decide whether to override the
given `:mode`. However mix is a build tool and shouldn't be used at
run-time. We are in the process of swapping SketchQL to using `mix
release` rather than `distillery`, which revealed our run-time
dependency on `mix`.

This commit moves `:mode_override` to be Application config, so that
SketchQL can set `config :trolleybus, mode_override: :full_sync` in
`config/test.exs`:

- https://github.com/sketch-hq/sketchql/pull/4774

Note, the SketchQL PR above should be merged _before_ this one to avoid
race conditions in SketchQL's test suite.